@@ -1,36 +1,46 @@
 {
     "name": "file-tree-alternative",
-    "version": "2.3.3",
+    "version": "2.6.0",
     "description": "This plugin allows you to have an alternative file tree view.",
     "main": "main.js",
     "scripts": {
-        "dev": "cross-env BUILD=dev obsidian-plugin dev src/main.ts",
-        "build": "cross-env BUILD=production obsidian-plugin build src/main.ts"
+        "dev": "rollup -c -w --bundleConfigAsCjs",
+        "build": "NODE_ENV=production rollup -c --bundleConfigAsCjs"
     },
     "keywords": [],
     "author": "",
     "license": "MIT",
     "devDependencies": {
-        "@types/common-tags": "^1.8.1",
+        "@babel/core": "^7.23.3",
+        "@babel/preset-react": "^7.23.3",
+        "@babel/preset-typescript": "^7.23.3",
+        "@rollup/plugin-babel": "^6.0.4",
+        "@rollup/plugin-commonjs": "^25.0.7",
+        "@rollup/plugin-json": "^6.0.1",
+        "@rollup/plugin-node-resolve": "^15.2.3",
+        "@rollup/plugin-replace": "^5.0.5",
+        "@rollup/plugin-typescript": "^11.1.5",
+        "@types/common-tags": "1.8.2",
         "@types/node": "^14.14.37",
-        "@types/react": "17.0.2",
-        "@types/react-dom": "17.0.2",
-        "cross-env": "^7",
-        "obsidian": "^0.12.0",
-        "obsidian-plugin-cli": "^0.8.0",
+        "@types/react": "18.2.0",
+        "@types/react-dom": "18.2.0",
+        "babel": "^6.23.0",
+        "million": "^2.6.4",
+        "obsidian": "^1.4.11",
         "prop-types": "^15.7.2",
+        "rollup": "^4.5.2",
         "tslib": "^2.2.0",
         "typescript": "^4.2.4"
     },
     "dependencies": {
-        "@ozntel/local-storage-handler": "^1.0.3",
-        "common-tags": "^1.8.2",
-        "dayjs": "^1.10.7",
-        "preact": "^10",
-        "react": "npm:@preact/compat@17.0.2",
-        "react-dom": "npm:@preact/compat@17.0.2",
-        "react-dropzone": "^11.3.2",
-        "react-icons": "^4.3.1",
-        "recoil": "^0.4.1"
+        "@ozntel/local-storage-handler": "1.0.3",
+        "common-tags": "1.8.2",
+        "dayjs": "1.11.10",
+        "react": "18.2.0",
+        "react-dom": "18.2.0",
+        "react-dropzone": "14.2.3",
+        "react-icons": "4.12.0",
+        "react-lazy-load": "^4.0.1",
+        "recoil": "0.7.7"
     }
 }
