@@ -1,20 +1,14 @@
 import { ItemView, WorkspaceLeaf } from 'obsidian';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { Root, createRoot } from 'react-dom/client';
 import FileTreeAlternativePlugin from './main';
 import MainTreeComponent from './components/MainView/MainComponent';
 import { RecoilRoot } from 'recoil';
 
-export const VIEW_TYPE = 'file-tree-view';
-export const VIEW_DISPLAY_TEXT = 'File Tree';
-export const ICON = 'sheets-in-box';
-
 export class FileTreeView extends ItemView {
     plugin: FileTreeAlternativePlugin;
     currentFolderPath: string;
-
-    // @ Temp Fix for Opening New File
-    navigation = false;
+    root: Root;
 
     constructor(leaf: WorkspaceLeaf, plugin: FileTreeAlternativePlugin) {
         super(leaf);
@@ -22,15 +16,15 @@ export class FileTreeView extends ItemView {
     }
 
     getViewType(): string {
-        return VIEW_TYPE;
+        return this.plugin?.VIEW_TYPE;
     }
 
     getDisplayText(): string {
-        return VIEW_DISPLAY_TEXT;
+        return this.plugin?.VIEW_DISPLAY_TEXT;
     }
 
     getIcon(): string {
-        return ICON;
+        return this.plugin?.ICON;
     }
 
     async onClose() {
@@ -38,7 +32,7 @@ export class FileTreeView extends ItemView {
     }
 
     destroy() {
-        ReactDOM.unmountComponentAtNode(this.contentEl);
+        if (this.root) this.root.unmount();
     }
 
     async onOpen(): Promise<void> {
@@ -48,13 +42,13 @@ export class FileTreeView extends ItemView {
 
     constructFileTree(folderPath: string, vaultChange: string) {
         this.destroy();
-        ReactDOM.render(
+        this.root = createRoot(this.contentEl);
+        this.root.render(
             <div className="file-tree-plugin-view">
                 <RecoilRoot>
                     <MainTreeComponent fileTreeView={this} plugin={this.plugin} />
                 </RecoilRoot>
-            </div>,
-            this.contentEl
+            </div>
         );
     }
 }
