@@ -11,6 +11,7 @@ type TreeProps = {
     open?: boolean;
     content?: string;
     onClick?: Function;
+    onDoubleClick: Function;
     onContextMenu?: Function;
     type?: any;
     style?: any;
@@ -75,8 +76,13 @@ export default function Tree(props: TreeProps) {
     };
     const folderContextMenuEvent = () => props.onContextMenu();
 
+    const folderNameDoubleClickEvent = (ev: React.MouseEvent) => {
+        setOpen(true);
+        props.onDoubleClick();
+    };
+
     // --> Icon
-    const Icon = useMemo(() => getFolderIcon(props.plugin, props.children, open), [open, props.children]);
+    const Icon = useMemo(() => getFolderIcon(props.plugin, props.children, open), [open, props.children, props.plugin.settings.folderIcon]);
 
     // --> Folder Count Map
     const folderCount = folderFileCountMap[props.folder.path];
@@ -118,6 +124,10 @@ export default function Tree(props: TreeProps) {
 
     const onFolderDragStart = (e: React.DragEvent<HTMLDivElement>, folder: TFolder) => {
         e.dataTransfer.setData('application/json', JSON.stringify({ folderPath: folder.path }));
+
+        let dragManager = (props.plugin.app as any).dragManager;
+        const dragData = dragManager.dragFile(e.nativeEvent, folder);
+        dragManager.onDragStart(e.nativeEvent, dragData);
     };
 
     return (
@@ -152,6 +162,7 @@ export default function Tree(props: TreeProps) {
                                 <div
                                     className="oz-folder-block"
                                     onClick={folderNameClickEvent}
+                                    onDoubleClick={folderNameDoubleClickEvent}
                                     onContextMenu={folderContextMenuEvent}
                                     {...longPressEvents}>
                                     <div className="oz-folder-type" style={{ marginRight: props.type ? 10 : 0 }}>
